+++ source/FileOperations/fileops.h
@@ -37,8 +37,8 @@ extern "C" {
 bool CreateSubfolder(const char * fullpath);
 bool CheckFile(const char * filepath);
 u64 FileSize(const char * filepath);
-int LoadFileToMem(const char * filepath, u8 **buffer, u64 *size);
-int LoadFileToMemWithProgress(const char *progressText, const char *filePath, u8 **buffer, u64 *size);
+int LoadFileToMem(const char * filepath, u8 **buffer, u32 *size);
+int LoadFileToMemWithProgress(const char *progressText, const char *filePath, u8 **buffer, u32 *size);
 int CopyFile(const char * src, const char * dest);
 int MoveFile(const char *srcpath, char *destdir);
 bool RenameFile(const char * srcpath, const char * destpath);
