+++ source/GUI/gui_imagedata.cpp
@@ -42,7 +42,7 @@ GuiImageData::GuiImageData(const char * filepath)
 	format = GX_TF_RGBA8;
 
 	u8 *buffer = NULL;
-	u64 size = 0;
+	u32 size = 0;
 
 	if(LoadFileToMem(filepath, &buffer, &size) < 0)
 		return;
